Move utils and types
