+++ packages/builder/src/stores/builder/components.ts
@@ -32,6 +32,8 @@ import {
 import { BudiStore } from "../BudiStore"
 import { Utils } from "@budibase/frontend-core"
 import {
+  ComponentDefinition,
+  ComponentSetting,
   Component as ComponentType,
   FieldType,
   Screen,
@@ -51,30 +53,6 @@ export interface ComponentState {
   componentToPaste?: Component
   settingsCache: Record<string, ComponentSetting[]>
   selectedScreenId?: string | null
-}
-
-export interface ComponentDefinition {
-  component: string
-  name: string
-  friendlyName?: string
-  hasChildren?: boolean
-  settings?: ComponentSetting[]
-  features?: Record<string, boolean>
-  typeSupportPresets?: Record<string, any>
-  legalDirectChildren: string[]
-  requiredAncestors?: string[]
-  illegalChildren: string[]
-}
-
-export interface ComponentSetting {
-  key: string
-  type: string
-  section?: string
-  name?: string
-  defaultValue?: any
-  selectAllFields?: boolean
-  resetOn?: string | string[]
-  settings?: ComponentSetting[]
 }
 
 export const INITIAL_COMPONENTS_STATE: ComponentState = {
